change data type of product_price to DOUBLE

+++ migrations/migrate/sqls/20210507222504-documents-up.sql
@@ -36,9 +36,9 @@ CREATE TABLE `documents` (
 CREATE TABLE `documents_products` (
   `document_id` INT NOT NULL,
   `product_id` INT NOT NULL,
-  `product_price` INT NOT NULL,
+  `product_price` DOUBLE NOT NULL,
   `product_quantity` INT NOT NULL,
-  `product_return_cost` INT,
+  `product_return_cost` DOUBLE,
   `tax_fee` DECIMAL(5,2) NOT NULL,
   `unit_tax_amount` DOUBLE NOT NULL,
   `discount_percentage` DECIMAL(5,2) DEFAULT NULL,
